changed int to double

+++ Test DriveTrain Team 1/src/main/java/frc/robot/Constants.java
@@ -21,7 +21,7 @@ public final class Constants {
         public static final int drive_lb = 13;
         public static final int drive_rf = 23;
         public static final int drive_rb = 15;
-        public static final int AutoSpeed = 0.5;
+        public static final double AutoSpeed = 0.5;
   }
 
 }
